SpreadsheetViewportNavigationList.parse was-SpreadsheetViewportNavigation.parse
- identical only return type is SpreadsheetViewportNavigationList was List<SpreadsheetViewportNavigation>.
