+++ src/main/java/walkingkooka/spreadsheet/reference/SpreadsheetViewportNavigation.java
@@ -17,43 +17,16 @@
 
 package walkingkooka.spreadsheet.reference;
 
-import walkingkooka.InvalidCharacterException;
 import walkingkooka.collect.list.Lists;
-import walkingkooka.datetime.DateTimeContexts;
-import walkingkooka.math.DecimalNumberContexts;
-import walkingkooka.predicate.character.CharPredicate;
-import walkingkooka.predicate.character.CharPredicates;
 import walkingkooka.spreadsheet.SpreadsheetViewportRectangle;
 import walkingkooka.spreadsheet.SpreadsheetViewportWindows;
 import walkingkooka.spreadsheet.SpreadsheetViewportWindowsFunction;
-import walkingkooka.spreadsheet.parser.SpreadsheetCellReferenceParserToken;
-import walkingkooka.spreadsheet.parser.SpreadsheetColumnReferenceParserToken;
-import walkingkooka.spreadsheet.parser.SpreadsheetParserContext;
-import walkingkooka.spreadsheet.parser.SpreadsheetParserContexts;
-import walkingkooka.spreadsheet.parser.SpreadsheetParserToken;
-import walkingkooka.spreadsheet.parser.SpreadsheetParsers;
-import walkingkooka.spreadsheet.parser.SpreadsheetRowReferenceParserToken;
-import walkingkooka.text.CaseSensitivity;
 import walkingkooka.text.HasText;
-import walkingkooka.text.cursor.TextCursor;
-import walkingkooka.text.cursor.TextCursors;
-import walkingkooka.text.cursor.parser.LongParserToken;
-import walkingkooka.text.cursor.parser.Parser;
-import walkingkooka.text.cursor.parser.ParserContext;
-import walkingkooka.text.cursor.parser.ParserReporters;
-import walkingkooka.text.cursor.parser.ParserToken;
-import walkingkooka.text.cursor.parser.Parsers;
-import walkingkooka.tree.expression.ExpressionNumberContexts;
-import walkingkooka.tree.expression.ExpressionNumberKind;
 
-import java.math.MathContext;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.function.Function;
-import java.util.function.IntFunction;
-import java.util.function.Supplier;
 
 /**
  * Captures a users input movement relative to another selection with a viewport
@@ -321,343 +294,6 @@ public abstract class SpreadsheetViewportNavigation implements HasText {
     @Override
     public final String toString() {
         return this.text();
-    }
-
-    /**
-     * Accepts text that has a more pretty form of any {@link SpreadsheetViewportNavigation enum value}.
-     * The text is identical to the enum name but in lower case and underscore replaced with dash.
-     * <br>
-     * {@link #extendLeftColumn()} = <pre>extend-left</pre>.
-     */
-    public static List<SpreadsheetViewportNavigation> parse(final String text) {
-        final TextCursor cursor = TextCursors.charSequence(text);
-        final List<SpreadsheetViewportNavigation> navigations = Lists.array();
-
-        final Supplier<IllegalArgumentException> ice = () -> new InvalidCharacterException(
-                text,
-                cursor.lineInfo()
-                        .textOffset()
-        );
-
-        while (false == cursor.isEmpty()) {
-            final SpreadsheetViewportNavigation navigation;
-
-            if (isMatch(LEFT, cursor)) {
-                navigation = parseSpaceColorRowOrPixels(
-                        cursor,
-                        COLUMN,
-                        SpreadsheetViewportNavigation::leftColumn,
-                        SpreadsheetViewportNavigation::leftPixel
-                );
-            } else {
-                if (isMatch(RIGHT, cursor)) {
-                    navigation = parseSpaceColorRowOrPixels(
-                            cursor,
-                            COLUMN,
-                            SpreadsheetViewportNavigation::rightColumn,
-                            SpreadsheetViewportNavigation::rightPixel
-                    );
-                } else {
-                    if (isMatch(UP, cursor)) {
-                        navigation = parseSpaceColorRowOrPixels(
-                                cursor,
-                                ROW,
-                                SpreadsheetViewportNavigation::upRow,
-                                SpreadsheetViewportNavigation::upPixel
-                        );
-                    } else {
-                        if (isMatch(DOWN, cursor)) {
-                            navigation = parseSpaceColorRowOrPixels(
-                                    cursor,
-                                    ROW,
-                                    SpreadsheetViewportNavigation::downRow,
-                                    SpreadsheetViewportNavigation::downPixel
-                            );
-                        } else {
-                            if (isMatch(EXTEND_LEFT, cursor)) {
-                                navigation = parseSpaceColorRowOrPixels(
-                                        cursor,
-                                        COLUMN,
-                                        SpreadsheetViewportNavigation::extendLeftColumn,
-                                        SpreadsheetViewportNavigation::extendLeftPixel
-                                );
-                            } else {
-                                if (isMatch(EXTEND_RIGHT, cursor)) {
-                                    navigation = parseSpaceColorRowOrPixels(
-                                            cursor,
-                                            COLUMN,
-                                            SpreadsheetViewportNavigation::extendRightColumn,
-                                            SpreadsheetViewportNavigation::extendRightPixel
-                                    );
-                                } else {
-                                    if (isMatch(EXTEND_UP, cursor)) {
-                                        navigation = parseSpaceColorRowOrPixels(
-                                                cursor,
-                                                ROW,
-                                                SpreadsheetViewportNavigation::extendUpRow,
-                                                SpreadsheetViewportNavigation::extendUpPixel
-                                        );
-                                    } else {
-                                        if (isMatch(EXTEND_DOWN, cursor)) {
-                                            navigation = parseSpaceColorRowOrPixels(
-                                                    cursor,
-                                                    ROW,
-                                                    SpreadsheetViewportNavigation::extendDownRow,
-                                                    SpreadsheetViewportNavigation::extendDownPixel
-                                            );
-                                        } else {
-                                            if (isMatch(SELECT, cursor)) {
-                                                navigation = parseCellColumnOrRow(
-                                                        cursor,
-                                                        SpreadsheetViewportNavigation::cell,
-                                                        SpreadsheetViewportNavigation::column,
-                                                        SpreadsheetViewportNavigation::row,
-                                                        ice
-                                                );
-                                            } else {
-                                                if (isMatch(EXTEND, cursor)) {
-                                                    navigation = parseCellColumnOrRow(
-                                                            cursor,
-                                                            SpreadsheetViewportNavigation::extendCell,
-                                                            SpreadsheetViewportNavigation::extendColumn,
-                                                            SpreadsheetViewportNavigation::extendRow,
-                                                            ice
-                                                    );
-                                                } else {
-                                                    throw new InvalidCharacterException(
-                                                            text,
-                                                            cursor.lineInfo()
-                                                                    .textOffset()
-                                                    );
-                                                }
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-
-            navigations.add(navigation);
-
-            if (cursor.isEmpty()) {
-                break;
-            }
-
-            SEPARATOR.parse(cursor, PARSER_CONTEXT);
-        }
-
-        return Lists.immutable(navigations);
-    }
-
-    private final static Parser<ParserContext> LEFT = stringParser("left");
-
-    private final static Parser<ParserContext> RIGHT = stringParser("right");
-
-    private final static Parser<ParserContext> UP = stringParser("up");
-
-    private final static Parser<ParserContext> DOWN = stringParser("down");
-
-    private final static Parser<ParserContext> EXTEND_LEFT = stringParser("extend-left");
-
-    private final static Parser<ParserContext> EXTEND_RIGHT = stringParser("extend-right");
-
-    private final static Parser<ParserContext> EXTEND_UP = stringParser("extend-up");
-
-    private final static Parser<ParserContext> EXTEND_DOWN = stringParser("extend-down");
-
-    private final static Parser<ParserContext> SEPARATOR = characterParserOrReport(
-            CharPredicates.is(SpreadsheetViewport.SEPARATOR.character())
-    );
-
-    /**
-     * Returns true if the parser was successful in matching text. The result is ignored.
-     */
-    private static boolean isMatch(final Parser<ParserContext> parser,
-                                   final TextCursor cursor) {
-        return parser.parse(
-                cursor,
-                PARSER_CONTEXT).isPresent();
-    }
-
-    /**
-     * Attempts to match the column or row suffix and if that fails expects a pixel value followed by px.
-     * <pre>
-     * left column
-     * left 123px
-     * </pre>
-     */
-    private static SpreadsheetViewportNavigation parseSpaceColorRowOrPixels(final TextCursor cursor,
-                                                                            final Parser<ParserContext> columnOrRowParser,
-                                                                            final Supplier<SpreadsheetViewportNavigation> columnOrRowNavigation,
-                                                                            final IntFunction<SpreadsheetViewportNavigation> columnOrRowPixel) {
-        parseSpace(cursor);
-
-        final SpreadsheetViewportNavigation navigation;
-
-        final Optional<ParserToken> maybeColumnOrRow = columnOrRowParser.parse(cursor, PARSER_CONTEXT);
-        if (maybeColumnOrRow.isPresent()) {
-            navigation = columnOrRowNavigation.get();
-        } else {
-            navigation = columnOrRowPixel.apply(
-                    VALUE.parse(
-                                    cursor,
-                                    PARSER_CONTEXT
-                            ).get()
-                            .cast(LongParserToken.class)
-                            .value()
-                            .intValue()
-            );
-
-            PX.parse(
-                    cursor,
-                    PARSER_CONTEXT
-            );
-        }
-
-        return navigation;
-    }
-
-    private final static Parser<ParserContext> VALUE = Parsers.longParser(10)
-            .orReport(ParserReporters.invalidCharacterException());
-
-    private final static Parser<ParserContext> CELL = stringParser("cell");
-
-    private final static Parser<ParserContext> COLUMN = stringParser("column");
-
-    private final static Parser<ParserContext> ROW = stringParser("row");
-
-    private final static Parser<ParserContext> PX = stringParser("px")
-            .orReport(ParserReporters.invalidCharacterException());
-
-
-    // select cell A1
-    // select column A
-    // select row B
-    private final static Parser<ParserContext> SELECT = stringParser("select");
-
-    // extend cell A1
-    // extend column A
-    // extend row B
-    private final static Parser<ParserContext> EXTEND = stringParser("extend");
-
-
-    // select cell A1
-    // select column AB
-    // select row 23
-    // extend cell A1
-    // extend column A
-    // extend row B
-    private static SpreadsheetViewportNavigation parseCellColumnOrRow(final TextCursor cursor,
-                                                                      final Function<SpreadsheetCellReference, SpreadsheetViewportNavigation> cell,
-                                                                      final Function<SpreadsheetColumnReference, SpreadsheetViewportNavigation> column,
-                                                                      final Function<SpreadsheetRowReference, SpreadsheetViewportNavigation> row,
-                                                                      final Supplier<IllegalArgumentException> invalidCharacter) {
-        parseSpace(cursor);
-
-        final SpreadsheetViewportNavigation navigation;
-
-        if (isMatch(CELL, cursor)) {
-            parseSpace(cursor);
-
-            navigation = cell.apply(
-                    parseSelection(
-                            CELL_PARSER,
-                            cursor,
-                            SpreadsheetCellReferenceParserToken.class
-                    ).reference()
-            );
-        } else {
-            if (isMatch(COLUMN, cursor)) {
-                parseSpace(cursor);
-
-                navigation = column.apply(
-                        parseSelection(
-                                COLUMN_PARSER,
-                                cursor,
-                                SpreadsheetColumnReferenceParserToken.class
-                        ).reference()
-                );
-            } else {
-                if (isMatch(ROW, cursor)) {
-                    parseSpace(cursor);
-
-                    navigation = row.apply(
-                            parseSelection(
-                                    ROW_PARSER,
-                                    cursor,
-                                    SpreadsheetRowReferenceParserToken.class
-                            ).reference()
-                    );
-                } else {
-                    throw invalidCharacter.get();
-                }
-            }
-        }
-
-        return navigation;
-    }
-
-    private final static Parser<ParserContext> CELL_PARSER = SpreadsheetParsers.cell()
-            .orReport(ParserReporters.invalidCharacterException())
-            .cast();
-
-    private final static Parser<ParserContext> COLUMN_PARSER = SpreadsheetParsers.column()
-            .orReport(ParserReporters.invalidCharacterException())
-            .cast();
-
-    private final static Parser<ParserContext> ROW_PARSER = SpreadsheetParsers.row()
-            .orReport(ParserReporters.invalidCharacterException())
-            .cast();
-
-    private static <T extends SpreadsheetParserToken> T parseSelection(final Parser<ParserContext> parser,
-                                                                       final TextCursor cursor,
-                                                                       final Class<T> parserToken) {
-        return parser.parse(
-                        cursor,
-                        PARSER_CONTEXT
-                ).get()
-                .cast(parserToken);
-    }
-
-    /**
-     * Parses a required space, throwing an exception if it was not found.
-     */
-    private static void parseSpace(final TextCursor cursor) {
-        SPACE.parse(cursor, PARSER_CONTEXT);
-    }
-
-    private final static Parser<ParserContext> SPACE = characterParserOrReport(
-            CharPredicates.is(' ')
-    );
-
-    private final static SpreadsheetParserContext PARSER_CONTEXT = SpreadsheetParserContexts.basic(
-            DateTimeContexts.fake(),
-            ExpressionNumberContexts.basic(
-                    ExpressionNumberKind.BIG_DECIMAL,
-                    DecimalNumberContexts.american(MathContext.DECIMAL32)
-            ),
-            ',' // value separator char
-    );
-
-    /**
-     * Factory that creates a {@link Parsers#string(String, CaseSensitivity)}.
-     */
-    private static Parser<ParserContext> characterParserOrReport(final CharPredicate predicate) {
-        return Parsers.character(predicate)
-                .orReport(ParserReporters.invalidCharacterException());
-    }
-
-    /**
-     * Factory that creates a {@link Parsers#string(String, CaseSensitivity)}.
-     */
-    private static Parser<ParserContext> stringParser(final String token) {
-        return Parsers.string(
-                token,
-                CaseSensitivity.SENSITIVE
-        );
     }
 
     /**
