+++ src/test/java/walkingkooka/spreadsheet/reference/SpreadsheetViewportNavigationListTest.java
@@ -18,13 +18,21 @@
 package walkingkooka.spreadsheet.reference;
 
 import org.junit.jupiter.api.Test;
+import walkingkooka.InvalidCharacterException;
 import walkingkooka.collect.list.ImmutableListTesting;
 import walkingkooka.collect.list.Lists;
+import walkingkooka.reflect.ClassTesting;
+import walkingkooka.reflect.JavaVisibility;
+import walkingkooka.test.ParseStringTesting;
+import walkingkooka.text.printer.TreePrintableTesting;
 
 import static org.junit.jupiter.api.Assertions.assertSame;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-public final class SpreadsheetViewportNavigationListTest implements ImmutableListTesting<SpreadsheetViewportNavigationList, SpreadsheetViewportNavigation> {
+public final class SpreadsheetViewportNavigationListTest implements ImmutableListTesting<SpreadsheetViewportNavigationList, SpreadsheetViewportNavigation>,
+        ParseStringTesting<SpreadsheetViewportNavigationList>,
+        TreePrintableTesting,
+        ClassTesting<SpreadsheetViewportNavigationList> {
 
     @Test
     public void testWithNullFails() {
@@ -43,13 +51,265 @@ public final class SpreadsheetViewportNavigationListTest implements ImmutableLis
         );
     }
 
+    // parse............................................................................................................
+
+    @Override
+    public void testParseStringEmptyFails() {
+        throw new UnsupportedOperationException();
+    }
+
+    @Test
+    public void testParseUnknownFails() {
+        this.parseStringFails(
+                "!invalid",
+                IllegalArgumentException.class
+        );
+    }
+
+    @Test
+    public void testParseUnknownFails2() {
+        final String text = "EXTEND-RIGHT";
+
+        this.parseStringFails(
+                text,
+                new InvalidCharacterException(
+                        text,
+                        0
+                )
+        );
+    }
+
+    @Test
+    public void testParseUnknownFails3() {
+        final String text = "right row";
+
+        this.parseStringFails(
+                text,
+                new InvalidCharacterException(
+                        text,
+                        "right ".length()
+                )
+        );
+    }
+
+    @Test
+    public void testParseInvalidValue() {
+        final String text = "right 123A";
+
+        this.parseStringFails(
+                text,
+                new InvalidCharacterException(
+                        text,
+                        "right 123".length()
+                )
+        );
+    }
+
+    @Test
+    public void testParseMissingPxSuffix() {
+        final String text = "down 123";
+
+        this.parseStringFails(
+                text,
+                new InvalidCharacterException(
+                        text,
+                        0
+                )
+        );
+    }
+
+    @Test
+    public void testParseIncompletePxSuffix() {
+        final String text = "extend-left 45p";
+
+        this.parseStringFails(
+                text,
+                new InvalidCharacterException(
+                        text,
+                        text.length() - 1
+                )
+        );
+    }
+
+    @Test
+    public void testParseEmpty() {
+        this.parseStringAndCheck(
+                ""
+        );
+    }
+
+    @Test
+    public void testParseLeftColumn() {
+        this.parseStringAndCheck0(
+                SpreadsheetViewportNavigation.leftColumn()
+        );
+    }
+
+    @Test
+    public void testParseRightColumn() {
+        this.parseStringAndCheck0(
+                SpreadsheetViewportNavigation.rightColumn()
+        );
+    }
+
+    @Test
+    public void testParseUpRow() {
+        this.parseStringAndCheck0(
+                SpreadsheetViewportNavigation.upRow()
+        );
+    }
+
+    @Test
+    public void testParseDownRow() {
+        this.parseStringAndCheck0(
+                SpreadsheetViewportNavigation.downRow()
+        );
+    }
+
+    @Test
+    public void testParseExtendLeftColumn() {
+        this.parseStringAndCheck0(
+                SpreadsheetViewportNavigation.extendLeftColumn()
+        );
+    }
+
+    @Test
+    public void testParseExtendRightColumn() {
+        this.parseStringAndCheck0(
+                SpreadsheetViewportNavigation.extendRightColumn()
+        );
+    }
+
+    @Test
+    public void testParseExtendUpRow() {
+        this.parseStringAndCheck0(
+                SpreadsheetViewportNavigation.extendUpRow()
+        );
+    }
+
+    @Test
+    public void testParseExtendDownRow() {
+        this.parseStringAndCheck0(
+                SpreadsheetViewportNavigation.extendDownRow()
+        );
+    }
+
+    @Test
+    public void testParseSelectCell() {
+        this.parseStringAndCheck0(
+                SpreadsheetViewportNavigation.cell(SpreadsheetSelection.A1)
+        );
+    }
+
+    @Test
+    public void testParseSelectCell2() {
+        this.parseStringAndCheck0(
+                SpreadsheetViewportNavigation.cell(SpreadsheetSelection.parseCell("B2"))
+        );
+    }
+
+    private void parseStringAndCheck0(final SpreadsheetViewportNavigation navigation) {
+        this.parseStringAndCheck(
+                navigation.text(),
+                navigation
+        );
+    }
+
+    @Test
+    public void testParseLeftColumnRightColumnUpRowDownRow() {
+        this.parseStringAndCheck(
+                "left column,right column,up row,down row",
+                SpreadsheetViewportNavigation.leftColumn(),
+                SpreadsheetViewportNavigation.rightColumn(),
+                SpreadsheetViewportNavigation.upRow(),
+                SpreadsheetViewportNavigation.downRow()
+        );
+    }
+
+    @Test
+    public void testParseExtendLeftColumnExtendRightColumnExtendUpRowExtendDownRow() {
+        this.parseStringAndCheck(
+                "extend-left column,extend-right column,extend-up row,extend-down row",
+                SpreadsheetViewportNavigation.extendLeftColumn(),
+                SpreadsheetViewportNavigation.extendRightColumn(),
+                SpreadsheetViewportNavigation.extendUpRow(),
+                SpreadsheetViewportNavigation.extendDownRow()
+        );
+    }
+
+    @Test
+    public void testParseLeftPixelRightPixelUpPixelDownPixel() {
+        this.parseStringAndCheck(
+                "left 10px,right 20px,up 30px,down 40px",
+                SpreadsheetViewportNavigation.leftPixel(10),
+                SpreadsheetViewportNavigation.rightPixel(20),
+                SpreadsheetViewportNavigation.upPixel(30),
+                SpreadsheetViewportNavigation.downPixel(40)
+        );
+    }
+
+    @Test
+    public void testParseExtendLeftPixelExtendRightPixelExtendUpPixelExtendDownPixel() {
+        this.parseStringAndCheck(
+                "extend-left 10px,extend-right 20px,extend-up 30px,extend-down 40px",
+                SpreadsheetViewportNavigation.extendLeftPixel(10),
+                SpreadsheetViewportNavigation.extendRightPixel(20),
+                SpreadsheetViewportNavigation.extendUpPixel(30),
+                SpreadsheetViewportNavigation.extendDownPixel(40)
+        );
+    }
+
+    @Test
+    public void testParseExtendLeftPixelExtendRightPixelExtendUpPixelExtendDownPixelSelectCell() {
+        this.parseStringAndCheck(
+                "extend-left 10px,extend-right 20px,extend-up 30px,extend-down 40px,select cell A1",
+                SpreadsheetViewportNavigation.extendLeftPixel(10),
+                SpreadsheetViewportNavigation.extendRightPixel(20),
+                SpreadsheetViewportNavigation.extendUpPixel(30),
+                SpreadsheetViewportNavigation.extendDownPixel(40),
+                SpreadsheetViewportNavigation.cell(SpreadsheetSelection.A1)
+        );
+    }
+
+    private void parseStringAndCheck(final String text,
+                                     final SpreadsheetViewportNavigation... navigations) {
+        this.parseStringAndCheck(
+                text,
+                SpreadsheetViewportNavigationList.with(
+                        Lists.of(navigations)
+                )
+        );
+    }
+
+    @Override
+    public SpreadsheetViewportNavigationList parseString(final String text) {
+        return SpreadsheetViewportNavigationList.parse(text);
+    }
+
+    @Override
+    public Class<? extends RuntimeException> parseStringFailedExpected(final Class<? extends RuntimeException> type) {
+        return type;
+    }
+
+    @Override
+    public RuntimeException parseStringFailedExpected(final RuntimeException cause) {
+        return cause;
+    }
+
     @Override
     public SpreadsheetViewportNavigationList createList() {
         return SpreadsheetViewportNavigationList.with(Lists.empty());
     }
 
+    // ClassTesting.....................................................................................................
+
     @Override
     public Class<SpreadsheetViewportNavigationList> type() {
         return SpreadsheetViewportNavigationList.class;
+    }
+
+    @Override
+    public JavaVisibility typeVisibility() {
+        return JavaVisibility.PUBLIC;
     }
 }
