+++ src/test/java/walkingkooka/spreadsheet/reference/SpreadsheetViewportNavigationTest.java
@@ -18,253 +18,17 @@
 package walkingkooka.spreadsheet.reference;
 
 import org.junit.jupiter.api.Test;
-import walkingkooka.InvalidCharacterException;
 import walkingkooka.collect.list.Lists;
 import walkingkooka.reflect.ClassTesting;
 import walkingkooka.reflect.JavaVisibility;
-import walkingkooka.test.ParseStringTesting;
 import walkingkooka.text.printer.TreePrintableTesting;
 
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-public final class SpreadsheetViewportNavigationTest implements ParseStringTesting<List<SpreadsheetViewportNavigation>>,
-        TreePrintableTesting,
+public final class SpreadsheetViewportNavigationTest implements TreePrintableTesting,
         ClassTesting<SpreadsheetViewportNavigation> {
-
-    // parse............................................................................................................
-
-    @Override
-    public void testParseStringEmptyFails() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Test
-    public void testParseUnknownFails() {
-        this.parseStringFails(
-                "!invalid",
-                IllegalArgumentException.class
-        );
-    }
-
-    @Test
-    public void testParseUnknownFails2() {
-        final String text = "EXTEND-RIGHT";
-
-        this.parseStringFails(
-                text,
-                new InvalidCharacterException(
-                        text,
-                        0
-                )
-        );
-    }
-
-    @Test
-    public void testParseUnknownFails3() {
-        final String text = "right row";
-
-        this.parseStringFails(
-                text,
-                new InvalidCharacterException(
-                        text,
-                        "right ".length()
-                )
-        );
-    }
-
-    @Test
-    public void testParseInvalidValue() {
-        final String text = "right 123A";
-
-        this.parseStringFails(
-                text,
-                new InvalidCharacterException(
-                        text,
-                        "right 123".length()
-                )
-        );
-    }
-
-    @Test
-    public void testParseMissingPxSuffix() {
-        final String text = "down 123";
-
-        this.parseStringFails(
-                text,
-                new InvalidCharacterException(
-                        text,
-                        0
-                )
-        );
-    }
-
-    @Test
-    public void testParseIncompletePxSuffix() {
-        final String text = "extend-left 45p";
-
-        this.parseStringFails(
-                text,
-                new InvalidCharacterException(
-                        text,
-                        text.length() - 1
-                )
-        );
-    }
-
-    @Test
-    public void testParseEmpty() {
-        this.parseStringAndCheck(
-                "",
-                Lists.empty()
-        );
-    }
-
-    @Test
-    public void testParseLeftColumn() {
-        this.parseStringAndCheck0(
-                SpreadsheetViewportNavigation.leftColumn()
-        );
-    }
-
-    @Test
-    public void testParseRightColumn() {
-        this.parseStringAndCheck0(
-                SpreadsheetViewportNavigation.rightColumn()
-        );
-    }
-
-    @Test
-    public void testParseUpRow() {
-        this.parseStringAndCheck0(
-                SpreadsheetViewportNavigation.upRow()
-        );
-    }
-
-    @Test
-    public void testParseDownRow() {
-        this.parseStringAndCheck0(
-                SpreadsheetViewportNavigation.downRow()
-        );
-    }
-
-    @Test
-    public void testParseExtendLeftColumn() {
-        this.parseStringAndCheck0(
-                SpreadsheetViewportNavigation.extendLeftColumn()
-        );
-    }
-
-    @Test
-    public void testParseExtendRightColumn() {
-        this.parseStringAndCheck0(
-                SpreadsheetViewportNavigation.extendRightColumn()
-        );
-    }
-
-    @Test
-    public void testParseExtendUpRow() {
-        this.parseStringAndCheck0(
-                SpreadsheetViewportNavigation.extendUpRow()
-        );
-    }
-
-    @Test
-    public void testParseExtendDownRow() {
-        this.parseStringAndCheck0(
-                SpreadsheetViewportNavigation.extendDownRow()
-        );
-    }
-
-    @Test
-    public void testParseSelectCell() {
-        this.parseStringAndCheck0(
-                SpreadsheetViewportNavigation.cell(SpreadsheetSelection.A1)
-        );
-    }
-
-    @Test
-    public void testParseSelectCell2() {
-        this.parseStringAndCheck0(
-                SpreadsheetViewportNavigation.cell(SpreadsheetSelection.parseCell("B2"))
-        );
-    }
-
-    private void parseStringAndCheck0(final SpreadsheetViewportNavigation navigation) {
-        this.parseStringAndCheck(
-                navigation.text(),
-                Lists.of(
-                        navigation
-                )
-        );
-    }
-
-    @Test
-    public void testParseLeftColumnRightColumnUpRowDownRow() {
-        this.parseStringAndCheck(
-                "left column,right column,up row,down row",
-                Lists.of(
-                        SpreadsheetViewportNavigation.leftColumn(),
-                        SpreadsheetViewportNavigation.rightColumn(),
-                        SpreadsheetViewportNavigation.upRow(),
-                        SpreadsheetViewportNavigation.downRow()
-                )
-        );
-    }
-
-    @Test
-    public void testParseExtendLeftColumnExtendRightColumnExtendUpRowExtendDownRow() {
-        this.parseStringAndCheck(
-                "extend-left column,extend-right column,extend-up row,extend-down row",
-                Lists.of(
-                        SpreadsheetViewportNavigation.extendLeftColumn(),
-                        SpreadsheetViewportNavigation.extendRightColumn(),
-                        SpreadsheetViewportNavigation.extendUpRow(),
-                        SpreadsheetViewportNavigation.extendDownRow()
-                )
-        );
-    }
-
-    @Test
-    public void testParseLeftPixelRightPixelUpPixelDownPixel() {
-        this.parseStringAndCheck(
-                "left 10px,right 20px,up 30px,down 40px",
-                Lists.of(
-                        SpreadsheetViewportNavigation.leftPixel(10),
-                        SpreadsheetViewportNavigation.rightPixel(20),
-                        SpreadsheetViewportNavigation.upPixel(30),
-                        SpreadsheetViewportNavigation.downPixel(40)
-                )
-        );
-    }
-
-    @Test
-    public void testParseExtendLeftPixelExtendRightPixelExtendUpPixelExtendDownPixel() {
-        this.parseStringAndCheck(
-                "extend-left 10px,extend-right 20px,extend-up 30px,extend-down 40px",
-                Lists.of(
-                        SpreadsheetViewportNavigation.extendLeftPixel(10),
-                        SpreadsheetViewportNavigation.extendRightPixel(20),
-                        SpreadsheetViewportNavigation.extendUpPixel(30),
-                        SpreadsheetViewportNavigation.extendDownPixel(40)
-                )
-        );
-    }
-
-    @Test
-    public void testParseExtendLeftPixelExtendRightPixelExtendUpPixelExtendDownPixelSelectCell() {
-        this.parseStringAndCheck(
-                "extend-left 10px,extend-right 20px,extend-up 30px,extend-down 40px,select cell A1",
-                Lists.of(
-                        SpreadsheetViewportNavigation.extendLeftPixel(10),
-                        SpreadsheetViewportNavigation.extendRightPixel(20),
-                        SpreadsheetViewportNavigation.extendUpPixel(30),
-                        SpreadsheetViewportNavigation.extendDownPixel(40),
-                        SpreadsheetViewportNavigation.cell(SpreadsheetSelection.A1)
-                )
-        );
-    }
 
     // compact..........................................................................................................
 
@@ -656,23 +420,6 @@ public final class SpreadsheetViewportNavigationTest implements ParseStringTesti
                 SpreadsheetViewportNavigation.compact(in),
                 () -> "compact " + in
         );
-    }
-
-    // ParseStringTesting...............................................................................................
-
-    @Override
-    public List<SpreadsheetViewportNavigation> parseString(final String text) {
-        return SpreadsheetViewportNavigation.parse(text);
-    }
-
-    @Override
-    public Class<? extends RuntimeException> parseStringFailedExpected(final Class<? extends RuntimeException> type) {
-        return type;
-    }
-
-    @Override
-    public RuntimeException parseStringFailedExpected(final RuntimeException cause) {
-        return cause;
     }
 
     // ClassTesting.....................................................................................................
